requirements.txt
gensim==4.1.2
pandas==1.3.4
pytz==2021.3
python-dotenv==0.19.1
tweepy==4.1.0
vaderSentiment==3.3.2